implementado crearPlato en panel PlatillosPanel
tambien se agrego lo siguiente:
ahora el crearIngredinet y crearPlato primero consulta si existe con un metodo de comprobarX, regresa del Dao al SA y otra vez dependiendo del booleano vuelve a meterse al dao para crearl, eliminar o modicicar el ingrediente, en el caso de plato solo lo implemente en crear y eliminar(por alguna razon en actualizar plato no esta usando un booleano) pero en ingredinte si lo implemente en los 3 casos estos

+++ Call_eat_02/src/Negocio/TransferPlato.java
@@ -19,7 +19,7 @@ public class TransferPlato {
 
 	}
 
-	public TransferPlato(String id, String nombre, Map<String, Integer> ingredientes, int precio,
+	public TransferPlato(String id, String nombre, Map<String, Integer> ingredientes, double precio,
 			String categoria, String iconPath) {
 		this.id = id;
 		this.nombre = nombre;
